requirements.txt
asgiref==3.7.2
attrs==23.2.0
cfgv==3.4.0
coverage==7.4.3
distlib==0.3.8
Django==4.2
django-cors-headers==4.3.1
django-debug-toolbar==4.2.0
django-extensions==3.2.3
django-phonenumber-field==7.3.0
django-rest-passwordreset==1.4.0
django-stubs==4.2.7
django-stubs-ext==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.1
exceptiongroup==1.2.0
Faker==24.8.0
filelock==3.13.1
gunicorn==21.2.0
identify==2.5.33
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
markdown-it-py==3.0.0
mdurl==0.1.2
mypy==1.7.1
mypy-extensions==1.0.0
nodeenv==1.8.0
packaging==23.2
phonenumberslite==8.13.27
platformdirs==4.1.0
pluggy==1.4.0
pprintpp==0.4.0
pre-commit==3.6.0
psycopg2-binary==2.9.9
Pygments==2.17.2
PyJWT==2.8.0
pytest==8.0.2
pytest-clarity==1.0.1
pytest-cov==4.1.0
pytest-django==4.8.0
pytest-testdox==3.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2024.1
PyYAML==6.0.1
referencing==0.33.0
rest-framework-simplejwt==0.0.2
rich==13.7.1
rpds-py==0.18.0
ruff==0.3.0
setuptools==69.1.1
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
types-pytz==2024.1.0.20240203
types-PyYAML==6.0.12.12
typing_extensions==4.9.0
tzdata==2023.4
uritemplate==4.1.1
virtualenv==20.25.0
XlsxWriter==3.2.0
